Add option to split the baseline grid.
It's not uncommon for a design to call for bottom-anchored content, such
as buttons, snackbars, drawers, etc.  Such content should be
grid-aligned internally, but its grid should originate at the bottom of
the screen and may not align with the top-anchored grid.

This change adds an option to split the grid at a given point on the
screen. There will then be two grids, one originating at TOP and one at
BOTTOM. If the grids do not align, a "gutter" will be left between them
to indicate the odd space. The gutter will always be narrower than the
baseline grid size.